Clean up responsibilities
