+++ go/obscuronode/enclave/rollupchain/l1_blockchain.go
@@ -1,4 +1,4 @@
-package enclave
+package rollupchain
 
 import (
 	"os"
