+++ go/obscuronode/host/interfaces.go
@@ -32,6 +32,6 @@ type StatsCollector interface {
 	// L2Recalc - called when a node has to discard the speculative work built on top of the winner of the gossip round.
 	L2Recalc(id common.Address)
 	NewBlock(block *types.Block)
-	NewRollup(node common.Address, rollup *nodecommon.Rollup)
+	NewRollup(node common.Address, rollup *nodecommon.EncryptedRollup)
 	RollupWithMoreRecentProof()
 }
