+++ integration/simulation/validate_chain.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/obscuronet/obscuro-playground/go/obscuronode/enclave/evm"
+	"github.com/obscuronet/obscuro-playground/go/obscuronode/enclave/bridge"
 
 	"github.com/obscuronet/obscuro-playground/go/obscuronode/obscuroclient"
 
@@ -275,7 +275,7 @@ func checkBlockchainOfObscuroNode(
 	totalAmountInSystem := s.Stats.TotalDepositedAmount - totalSuccessfullyWithdrawn
 	total := uint64(0)
 	for _, wallet := range s.Params.Wallets.SimObsWallets {
-		total += balance(nodeClient, wallet.Address(), s.Params.Wallets.Tokens[evm.BTC].L2ContractAddress)
+		total += balance(nodeClient, wallet.Address(), s.Params.Wallets.Tokens[bridge.BTC].L2ContractAddress)
 	}
 
 	if total != totalAmountInSystem {
